Generalise the AdConfimationLetterService
- We will use this for lower tier registrations also
