+++ app/controllers/concerns/can_complete_if_possible.rb
@@ -10,7 +10,7 @@ module CanCompleteIfPossible
       return false unless can_complete?
 
       WasteCarriersEngine::RegistrationActivationService.run(registration: @resource)
-      Notify::UpperTierAdConfirmationLetterService.run(registration: @resource)
+      Notify::AdConfirmationLetterService.run(registration: @resource)
 
       true
     rescue StandardError => e
