+++ app/controllers/registration_conviction_approval_forms_controller.rb
@@ -50,7 +50,7 @@ class RegistrationConvictionApprovalFormsController < ApplicationController
     return if @registration.unpaid_balance?
 
     WasteCarriersEngine::RegistrationActivationService.run(registration: @registration)
-    Notify::UpperTierAdConfirmationLetterService.run(registration: @registration)
+    Notify::AdConfirmationLetterService.run(registration: @registration)
   rescue StandardError => e
     Airbrake.notify(e, reg_identifier: @registration.reg_identifier)
     Rails.logger.error e
